refactor : Change file name to store
