--- src/pages/review/Review.styles.ts
@@ -1,19 +1,0 @@
-import styled from 'styled-components'
-
-export const ReviewWrapper = styled.section`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 50px 0;
-`
-export const ReviewMapWrapper = styled.div`
-  margin-top: 20px;
-  width: 80%;
-  min-width: 1000px;
-  height: 500px;
-
-  .map {
-    width: 100%;
-    height: 100%;
-  }
-`
